Graph visibility, graph canvas dimensions modified and hyperlinks added to graph nodes.
