+++ gnowsys-ndf/gnowsys_ndf/ndf/templates/ndf/node_details_base.html
@@ -145,7 +145,10 @@ src="http://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS-MML_HTMLorM
 
   <dl class="row tabs" data-tab>
     <dd class="active"><a href="#view-page"><i class="fi-eye"></i> Read</a></dd>
-    <dd><a href="#view-graph" data-reveal-id="view-graph"><i class="fi-trees"></i> Explore</a></dd>
+      {% if '/details/' in request.path  or '/image_detail/' in request.path  or '/video_detail/' in request.path%}
+        <dd><a href="#view-graph" data-reveal-id="view-graph"><i class="fi-trees"></i> Explore</a></dd>
+      {% endif %}
+    
     <dd><a href="#view-comments"><i class="fi-comment"></i> Discuss</a></dd>
   </dl>
 
@@ -194,16 +197,14 @@ src="http://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS-MML_HTMLorM
        {% endwith %}
      </div>
 
+
      <!-- Tab View Graph -->
-
-
      <div class="content reveal-modal x-large graph-height" id="view-graph" data-reveal>    
-      <h3>Exlore Object</h3>
-      {% if '/details/' in request.path %}
-      {% include "ndf/node_graph.html" with graphData=graphData %}
+      <h3>{{node.name}}</h3>
+      {% if '/details/' in request.path  or '/image_detail/' in request.path  or '/video_detail/' in request.path%}
+        <a class="close-reveal-modal">&#215;</a>
+        {% include "ndf/node_graph.html" %}
       {% endif %}
-      <!-- </div> -->
-      <!-- </div> -->
     </div>
 
   </section>
