+++ gnowsys-ndf/gnowsys_ndf/ndf/templates/ndf/node_graph.html
@@ -66,17 +66,13 @@
 
 <body>
 
-	<div id="chart" data-graph="{{graphData}}" data-loc="{{request.META.PATH_INFO}}">
+	<div id="chart">
 	</div>
 
 	<script type="text/javascript">
 	
-	var loc = ($('#chart').attr("data-loc")).split("/")[3];
-
-	var jsonData = $.parseJSON($("#chart").attr("data-graph"));
-
-	// graphArea = $("#view-graph").height();
-	// alert(graphArea);
+	var graphHeight = ($(window).height())*0.55,
+		graphWidth  = ($(window).width())*0.75;
 
 	function init(a,b)
 	{
@@ -105,19 +101,19 @@
 		$(function fgraph() 
 		{	
 
-			// d3.json(jsonData, function (json1){ 
-				json1 = jsonData;
+			d3.json("{% url 'get_graph_json' group_name %}?id={{node.pk}}", function (json1){ 
+				// json1 = jsonData;
 				metadata=json1.node_metadata;
 				relations=json1.relations;
 
 				init(metadata,relations);
-				load(jsonData.node_metadata[0]._id) 
-			// });
+				load(json1.node_metadata[0]._id) 
+			});
 
 			function load(key)
 			{
-				var w = 550,
-				h = 300,
+				var w = graphWidth,
+				h = graphHeight,
 				fill = d3.scale.category20();
 
 				var vis = d3.select("#chart")
@@ -160,7 +156,7 @@
 			        	}
 			        	return acc;
 			        }, []);
-	        
+	        		
 		        force.nodes(nodes);
 		        force.links(edges);
 		        force.start();
@@ -197,7 +193,7 @@
 
                 var new_g = node.enter().append("svg:a")
                 .attr("class", function(d) { var e=(d._id).charAt(0); if (d._id==key) return "mainnode"; else if (e=="-") return "nodetext"; else if (isNaN(d._id)) return "relnode"  ; else return "node"; })
-                // .attr("xlink:href", function(d){ if(d.refType == 'relation') return d.url})
+                .attr("xlink:href", function(d){ if(d.refType == 'relation') return d.url})
 
                 .call(force.drag);
 
@@ -213,7 +209,6 @@
                 .append("svg:path")
                 .attr("d","M 0 0 L 10 5 L 0 10 z")
                 .attr("fill", "#6D6666"); 
-
 
 
                 text1 = new_g.append("svg:text")
@@ -253,7 +248,7 @@
                 .style("fill-opacity", ".2")
                 // .style("stroke", "#666")
                 // .style("stroke-width", "1.5px")
-                .style("fill", function(d) {if (d.refType=="Objecttype") return "blue"; else if(d.refType=="GSystem") return "#0000ff"; else return "none"});
+                .style("fill", function(d) {if (d.refType=="Objecttype") return "blue"; else if(d.refType=="GSystem") return "#008759"; else return "none"});
 
                 node.exit().remove();
