+++ gnowsys-ndf/gnowsys_ndf/ndf/urls/ajax-urls.py
@@ -9,7 +9,8 @@ urlpatterns = patterns('gnowsys_ndf.ndf.views.ajax_views',
                        url(r'^collection/', 'select_drawer', name='select_drawer'),
                        url(r'^change_group_settings/', 'change_group_settings', name='change_group_settings'),                 
                        url(r'^make_module/', 'make_module_set', name='make_module'),                 
-                       url(r'^get_module_json/', 'get_module_json', name='get_module_json')
+                       url(r'^get_module_json/', 'get_module_json', name='get_module_json'),
+                       url(r'^get_graph_json/', 'graph_nodes', name='get_graph_json')
                        
 
 )
